refactor: fix cannot deploy
