+++ src/pages/Homepage.tsx
@@ -2,7 +2,7 @@ import React from "react";
 import Content from "../components/Content";
 import Footer from "../components/Footer";
 import Header from "../components/Header";
-import Layout from "../components/Layout";
+import Layout from "../components/layoutDasar";
 import Navbar from "../components/Navbar";
 
 const Homepage = () => {
