+++ src/pages/aboutmeProfil.tsx
@@ -9,7 +9,7 @@ import Swal from "../utils/swal";
 import ButtonLogin from "../components/buttonLogin";
 import InputProfil from "../components/inputProfil";
 import SideNav from "../components/sideNav";
-import Layout from "../components/Layout";
+import Layout from "../components/layoutDasar";
 import Navbar from "../components/Navbar";
 
 const aboutmeProfil = () => {
