+++ src/pages/auth/login.tsx
@@ -9,7 +9,7 @@ import { handleAuth } from "../../utils/redux/reducer/reducer";
 
 import ButtonRegister from "../../components/buttonRegister";
 import ButtonLogin from "../../components/buttonLogin";
-import Layout from "../../components/Layout";
+import Layout from "../../components/layoutDasar";
 import Input from "../../components/input";
 
 import imgLogin from "../../assets/imgLogin.svg";
